+++ commands/check.go
@@ -1,11 +1,7 @@
 package commands
 
 import (
-	"fmt"
-
-	consulapi "github.com/hashicorp/consul/api"
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
 )
 
 func newCheckCommand() *cobra.Command {
@@ -25,234 +21,4 @@ func newCheckCommand() *cobra.Command {
 	cmd.AddCommand(newCheckWarnCommand())
 
 	return cmd
-}
-
-// Deregister functions
-
-func newCheckDeregisterCommand() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "deregister",
-		Short: "Remove a check from the agent",
-		Long:  "Remove a check from the agent",
-		RunE:  checkDeregister,
-	}
-
-	return cmd
-}
-
-func checkDeregister(cmd *cobra.Command, args []string) error {
-	if len(args) != 1 {
-		return fmt.Errorf("A single check id must be specified")
-	}
-	checkId := args[0]
-
-	viper.BindPFlags(cmd.Flags())
-
-	client, err := newAgent()
-	if err != nil {
-		return err
-	}
-
-	err = client.CheckDeregister(checkId)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// Fail functions
-
-func newCheckFailCommand() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "fail <checkId>",
-		Short: "Mark a local check as critical",
-		Long:  "Mark a local check as critical",
-		RunE:  checkFail,
-	}
-
-	cmd.Flags().String("note", "", "Message to associate with check status")
-
-	return cmd
-}
-
-func checkFail(cmd *cobra.Command, args []string) error {
-	if len(args) != 1 {
-		return fmt.Errorf("A single check id must be specified")
-	}
-	checkId := args[0]
-
-	viper.BindPFlags(cmd.Flags())
-
-	client, err := newAgent()
-	if err != nil {
-		return err
-	}
-
-	err = client.FailTTL(checkId, viper.GetString("note"))
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// Pass functions
-
-func newCheckPassCommand() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "pass <checkId>",
-		Short: "Mark a local check as passing",
-		Long:  "Mark a local check as passing",
-		RunE:  checkPass,
-	}
-
-	cmd.Flags().String("note", "", "Message to associate with check status")
-
-	return cmd
-}
-
-func checkPass(cmd *cobra.Command, args []string) error {
-	if len(args) != 1 {
-		return fmt.Errorf("A single check id must be specified")
-	}
-	checkId := args[0]
-
-	viper.BindPFlags(cmd.Flags())
-
-	client, err := newAgent()
-	if err != nil {
-		return err
-	}
-
-	err = client.PassTTL(checkId, viper.GetString("note"))
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// Register functions
-
-var registerLongHelp = `Register a new local check
-
-  If --id is not specified, the checkName is used. There cannot\
-be duplicate IDs per agent however.
-
-  Only one of --http, --script and --ttl can be specified.
-`
-
-func newCheckRegisterCommand() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "register <checkName>",
-		Short: "Register a new local check",
-		Long:  registerLongHelp,
-		RunE:  checkRegister,
-	}
-
-	addCheckOptions(cmd)
-	addRawOption(cmd)
-
-	cmd.Flags().String("service-id", "", "Service ID to associate check")
-
-	return cmd
-}
-
-func checkRegister(cmd *cobra.Command, args []string) error {
-	viper.BindPFlags(cmd.Flags())
-
-	var check consulapi.AgentCheckRegistration
-
-	if raw := viper.GetString("raw"); raw != "" {
-		if err := readRawJSON(raw, &check); err != nil {
-			return err
-		}
-	} else {
-		if len(args) != 1 {
-			return fmt.Errorf("A single check id must be specified")
-		}
-		checkName := args[0]
-
-		checkCount := 0
-		if viper.GetString("http") != "" {
-			checkCount = checkCount + 1
-		}
-		if viper.GetString("script") != "" {
-			checkCount = checkCount + 1
-		}
-		if viper.GetString("ttl") != "" {
-			checkCount = checkCount + 1
-		}
-
-		if checkCount > 1 {
-			return fmt.Errorf("Only one of --http, --script or --ttl can be specified")
-		}
-
-		check = consulapi.AgentCheckRegistration{
-			ID:        viper.GetString("id"),
-			Name:      checkName,
-			ServiceID: viper.GetString("service-id"),
-			Notes:     viper.GetString("notes"),
-			AgentServiceCheck: consulapi.AgentServiceCheck{
-				Script:            viper.GetString("script"),
-				HTTP:              viper.GetString("http"),
-				Interval:          viper.GetString("interval"),
-				TTL:               viper.GetString("ttl"),
-				TLSSkipVerify:     viper.GetBool("skip-verify"),
-				DockerContainerID: viper.GetString("docker-id"),
-				Shell:             viper.GetString("shell"),
-				Notes:             viper.GetString("notes"),
-				DeregisterCriticalServiceAfter: viper.GetString("deregister-crit"),
-			},
-		}
-	}
-
-	client, err := newAgent()
-	if err != nil {
-		return err
-	}
-
-	err = client.CheckRegister(&check)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// Warn functions
-
-func newCheckWarnCommand() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "warn <checkId>",
-		Short: "Mark a local check as warning",
-		Long:  "Mark a local check as warning",
-		RunE:  checkWarn,
-	}
-
-	cmd.Flags().String("note", "", "Message to associate with check status")
-
-	return cmd
-}
-
-func checkWarn(cmd *cobra.Command, args []string) error {
-	if len(args) != 1 {
-		return fmt.Errorf("A single check id must be specified")
-	}
-	checkId := args[0]
-
-	viper.BindPFlags(cmd.Flags())
-
-	client, err := newAgent()
-	if err != nil {
-		return err
-	}
-
-	err = client.WarnTTL(checkId, viper.GetString("note"))
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
