+++ commands/status.go
@@ -2,7 +2,6 @@ package commands
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
 )
 
 func newStatusCommand() *cobra.Command {
@@ -19,64 +18,4 @@ func newStatusCommand() *cobra.Command {
 	cmd.AddCommand(newStatusPeersCommand())
 
 	return cmd
-}
-
-func newStatusLeaderCommand() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "leader",
-		Short: "Get the current Raft leader",
-		Long:  "Get the current Raft leader",
-		RunE:  statusLeader,
-	}
-
-	addTemplateOption(cmd)
-
-	return cmd
-}
-
-func statusLeader(cmd *cobra.Command, args []string) error {
-	viper.BindPFlags(cmd.Flags())
-
-	client, err := newStatus()
-	if err != nil {
-		return err
-	}
-
-	l, err := client.Leader()
-	if err != nil {
-		return err
-	}
-
-	return output(l)
-}
-
-// Peers functions
-
-func newStatusPeersCommand() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "peers",
-		Short: "Get the current Raft peers",
-		Long:  "Get the current Raft peers",
-		RunE:  statusPeers,
-	}
-
-	addTemplateOption(cmd)
-
-	return cmd
-}
-
-func statusPeers(cmd *cobra.Command, args []string) error {
-	viper.BindPFlags(cmd.Flags())
-
-	client, err := newStatus()
-	if err != nil {
-		return err
-	}
-
-	l, err := client.Peers()
-	if err != nil {
-		return err
-	}
-
-	return output(l)
 }
